+++ source/fundamentals/data-formats/bson.txt
@@ -50,18 +50,8 @@ and values in the document. Each ``BsonElement`` can be either an instance of th
 ``BsonElement`` class or a field-value pair inside curly braces ( ``{}`` ).
 
 The following code sample shows how to create a ``BsonDocument`` object to represent the 
-example BSON document. Each ``BsonElement`` is an instance of the ``BsonElement``
-class.
-
-.. literalinclude:: ../../includes/fundamentals/code-examples/Bson.cs
-   :start-after: start-bson-elements
-   :end-before: end-bson-elements
-   :language: csharp 
-   :copyable:
-   :dedent:
-
-The following code sample creates the same ``BsonDocument``, but each ``BsonElement``
-is a field-value pair: 
+example BSON document. Each key-value pair in the ``BsonDocument`` object is a 
+``BsonElement`` object. 
 
 .. literalinclude:: ../../includes/fundamentals/code-examples/Bson.cs
    :start-after: start-create
@@ -69,11 +59,6 @@ is a field-value pair:
    :language: csharp 
    :copyable:
    :dedent:
-
-.. tip::
-   
-   Each pair of curly braces in a BSON document requires one new ``BsonDocument`` object
-   in {+language+}.
 
 Change a BSON Document
 ----------------------
